Create i18n folder in source and moved all dicts into it, i18n is more used for internationalization
