+++ front-end/src/i18n/dictionaries/en.json
